Add IP testing to GrafDashboard, refactored layout, added types imports for Dashboard.tsx
