+++ package.json
@@ -30,6 +30,7 @@
     "sunburst-chart": "^1.18.0"
   },
   "devDependencies": {
+    "@types/jest": "^29.5.0",
     "@types/node": "18.15.11",
     "@types/react": "18.0.35",
     "sass": "^1.62.0"
